+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -100,7 +100,7 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-09-11T00:37:36.779Z")
+    updated: Optional[datetime] = parse_datetime("2026-09-11T12:15:32.022Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")