+++ packages/orq-rc/RELEASES.md
@@ -10859,3 +10859,13 @@ Based on:
 - [python v4.15.0-rc.39] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.39] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.39 - packages/orq-rc
+
+## 2026-09-11 12:18:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.40] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.40] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.40 - packages/orq-rc